+++ src/Google/Api/Ads/Dfp/v201502/CreativeTemplateService.php
@@ -5,7 +5,7 @@
  *
  * PHP version 5
  *
- * Copyright 2014, Google Inc. All Rights Reserved.
+ * Copyright 2016, Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,7 +22,7 @@
  * @package    GoogleApiAdsDfp
  * @subpackage v201502
  * @category   WebServices
- * @copyright  2014, Google Inc. All Rights Reserved.
+ * @copyright  2016, Google Inc. All Rights Reserved.
  * @license    http://www.apache.org/licenses/LICENSE-2.0 Apache License,
  *             Version 2.0
  */
@@ -484,6 +484,50 @@ if (!class_exists("ListStringCreativeTemplateVariableVariableChoice", false)) {
     public function __construct($label = null, $value = null) {
       $this->label = $label;
       $this->value = $value;
+    }
+
+  }
+}
+
+if (!class_exists("CreativeTemplateOperationError", false)) {
+  /**
+   * An error that can occur while performing an operation on a creative template.
+   * @package GoogleApiAdsDfp
+   * @subpackage v201502
+   */
+  class CreativeTemplateOperationError extends ApiError {
+
+    const WSDL_NAMESPACE = "https://www.google.com/apis/ads/publisher/v201502";
+    const XSI_TYPE = "CreativeTemplateOperationError";
+
+    /**
+     * @access public
+     * @var tnsCreativeTemplateOperationErrorReason
+     */
+    public $reason;
+
+    /**
+     * Gets the namesapce of this class
+     * @return string the namespace of this class
+     */
+    public function getNamespace() {
+      return self::WSDL_NAMESPACE;
+    }
+
+    /**
+     * Gets the xsi:type name of this class
+     * @return string the xsi:type name of this class
+     */
+    public function getXsiTypeName() {
+      return self::XSI_TYPE;
+    }
+
+    public function __construct($reason = null, $fieldPath = null, $trigger = null, $errorString = null) {
+      parent::__construct();
+      $this->reason = $reason;
+      $this->fieldPath = $fieldPath;
+      $this->trigger = $trigger;
+      $this->errorString = $errorString;
     }
 
   }
@@ -1970,6 +2014,39 @@ if (!class_exists("CreativeTemplateErrorReason", false)) {
   }
 }
 
+if (!class_exists("CreativeTemplateOperationErrorReason", false)) {
+  /**
+   * The reasons for the target error.
+   * @package GoogleApiAdsDfp
+   * @subpackage v201502
+   */
+  class CreativeTemplateOperationErrorReason {
+
+    const WSDL_NAMESPACE = "https://www.google.com/apis/ads/publisher/v201502";
+    const XSI_TYPE = "CreativeTemplateOperationError.Reason";
+
+    /**
+     * Gets the namesapce of this class
+     * @return string the namespace of this class
+     */
+    public function getNamespace() {
+      return self::WSDL_NAMESPACE;
+    }
+
+    /**
+     * Gets the xsi:type name of this class
+     * @return string the xsi:type name of this class
+     */
+    public function getXsiTypeName() {
+      return self::XSI_TYPE;
+    }
+
+    public function __construct() {
+    }
+
+  }
+}
+
 if (!class_exists("CreativeTemplateStatus", false)) {
   /**
    * Describes status of the creative template
@@ -3332,6 +3409,7 @@ if (!class_exists("CreativeTemplateService", false)) {
       "ListStringCreativeTemplateVariable" => "ListStringCreativeTemplateVariable",
       "ListStringCreativeTemplateVariable.VariableChoice" => "ListStringCreativeTemplateVariableVariableChoice",
       "LongCreativeTemplateVariable" => "LongCreativeTemplateVariable",
+      "CreativeTemplateOperationError" => "CreativeTemplateOperationError",
       "CreativeTemplatePage" => "CreativeTemplatePage",
       "StringCreativeTemplateVariable" => "StringCreativeTemplateVariable",
       "UrlCreativeTemplateVariable" => "UrlCreativeTemplateVariable",
@@ -3372,6 +3450,7 @@ if (!class_exists("CreativeTemplateService", false)) {
       "CommonError.Reason" => "CommonErrorReason",
       "AssetCreativeTemplateVariable.MimeType" => "AssetCreativeTemplateVariableMimeType",
       "CreativeTemplateError.Reason" => "CreativeTemplateErrorReason",
+      "CreativeTemplateOperationError.Reason" => "CreativeTemplateOperationErrorReason",
       "CreativeTemplateStatus" => "CreativeTemplateStatus",
       "CreativeTemplateType" => "CreativeTemplateType",
       "FeatureError.Reason" => "FeatureErrorReason",